Restructuring navbar.component.tsx location
